+++ src/SMAPI/ITranslationHelper.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using StardewValley;
 
 namespace StardewModdingAPI
@@ -30,5 +30,10 @@ namespace StardewModdingAPI
         /// <param name="key">The translation key.</param>
         /// <param name="tokens">An object containing token key/value pairs. This can be an anonymous object (like <c>new { value = 42, name = "Cranberries" }</c>), a dictionary, or a class instance.</param>
         Translation Get(string key, object tokens);
+
+        /// <summary>Get a translation in every locale for which it's defined.</summary>
+        /// <param name="key">The translation key.</param>
+        /// <param name="withFallback">Whether to add duplicate translations for locale fallback. For example, if a translation is defined in <c>default.json</c> but not <c>fr.json</c>, setting this to true will add a <c>fr</c> entry which duplicates the default text.</param>
+        IDictionary<string, Translation> GetInAllLocales(string key, bool withFallback = false);
     }
 }